reorganise include/ layout and promote headers; remove legacy files
