+++ CMakeLists.txt
@@ -348,7 +348,6 @@ set(HYDROCHRONO_SOURCES
     src/chloadaddedmass.cpp
     src/hydro_forces.cpp
     src/helper.cpp
-    src/wave_types.cpp
     # Hydro modules
     src/hydro/waves/wave_base.cpp
     src/hydro/waves/wave_utilities.cpp
