--- include/hydroc/hydrochrono_runner/run_hydrochrono_from_yaml.h
@@ -1,8 +1,0 @@
-#pragma once
-
-namespace hydroc {
-
-/// Entry point for HydroChrono YAML simulation runner
-int RunHydroChronoFromYAML(int argc, char* argv[]);
-
-}  // namespace hydroc
