+++ include/hydroc/runner/run_from_yaml.h
@@ -1,0 +1,17 @@
+/*********************************************************************
+ * @file  run_from_yaml.h
+ * @brief Entry point for HydroChrono YAML simulation runner.
+ *********************************************************************/
+
+#ifndef HYDROC_RUNNER_RUN_FROM_YAML_H
+#define HYDROC_RUNNER_RUN_FROM_YAML_H
+
+namespace hydroc {
+
+/// Entry point for HydroChrono YAML simulation runner
+int RunHydroChronoFromYAML(int argc, char* argv[]);
+
+}  // namespace hydroc
+
+#endif  // HYDROC_RUNNER_RUN_FROM_YAML_H
+
