--- include/hydroc/wave_types.h
@@ -1,10 +1,0 @@
-
-#ifndef WAVE_TYPES_H
-#define WAVE_TYPES_H
-
-#include "../../src/hydro/waves/wave_utilities.h"
-#include "../../src/hydro/waves/wave_base.h"
-#include "../../src/hydro/waves/regular_wave.h"
-#include "../../src/hydro/waves/irregular_wave.h"
-
-#endif
