+++ src/hydro/force_components/hydrostatics_component.h
@@ -6,8 +6,8 @@
 #ifndef HYDRO_FORCE_COMPONENTS_HYDROSTATICS_COMPONENT_H
 #define HYDRO_FORCE_COMPONENTS_HYDROSTATICS_COMPONENT_H
 
-#include "../core/hydro_force_component.h"
-#include <hydroc/system_state.h>
+#include <hydroc/core/force_component.h>
+#include <hydroc/core/system_state.h>
 #include <Eigen/Dense>
 #include <vector>
 #include <memory>
