+++ src/hydrochrono_runner/run_hydrochrono_from_yaml.cpp
@@ -5,7 +5,9 @@
 #include "../hydro/config/yaml_parser.h"
 #include <hydroc/hydro_forces.h>
 #include <hydroc/io/simulation_export.h>
-#include <hydroc/wave_types.h>
+#include <hydroc/waves/wave_base.h>
+#include <hydroc/waves/regular_wave.h>
+#include <hydroc/waves/irregular_wave.h>
 
 #include <chrono_parsers/yaml/ChParserMbsYAML.h>
 #include <chrono/physics/ChSystem.h>
