--- src/wave_types.cpp
@@ -1,9 +1,0 @@
-/*********************************************************************
- * @file  wave_types.cpp
- * @brief Legacy translation unit kept for build compatibility.
- *
- * All wave implementations now live under src/hydro/waves/.
- *********************************************************************/
-
-#include <hydroc/wave_types.h>
-
